Allow runners to reject workflows based on a TOKEN
Allow runners to expect a variable in a workflow, ACTIONS_RUNNER_TOKEN
which is expected to match an environment variable specified to the
runner, GITHUB_ACTIONS_RUNNER_TOKEN.

Only workflows that specify the correct token will be allowed to run.

Workflows can use a secret, which gives finer-grained control over the
workflows that can run on a particular runner.